reverted change--was adding to the wrong file

+++ project/PPO_discrete.py
@@ -640,7 +640,7 @@ if __name__ == "__main__":
                                 end = start + args.minibatch_size
                                 mb_inds = b_inds[start:end]
 
-                                _, newlogprob, entropy, newvalue = agent_instance.get_action_and_value(b_obs[mb_inds], b_actions[mb_inds])
+                                _, newlogprob, entropy, newvalue = agent_instance.get_action_and_value(b_obs[mb_inds], b_actions.long()[mb_inds])
                                 logratio = newlogprob - b_logprobs[mb_inds]
                                 ratio = logratio.exp()
